+++ app/src/main/java/hous/release/android/presentation/todo/add/AddToDoVIewModel.kt
@@ -1,101 +1,30 @@
 package hous.release.android.presentation.todo.add
 
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import hous.release.android.presentation.our_rules.type.ButtonState
+import hous.release.android.presentation.todo.viewmodel.UpdateToDoUiState
+import hous.release.android.presentation.todo.viewmodel.UpdateToDoViewModel
 import hous.release.domain.entity.HomyType
 import hous.release.domain.entity.response.ToDoUser
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class AddToDoVIewModel @Inject constructor() : ViewModel() {
-    val todoName = MutableStateFlow("")
-    private var _uiState = MutableStateFlow(dummyUiState)
-    val uiState = _uiState.asStateFlow()
+class AddToDoVIewModel @Inject constructor() : UpdateToDoViewModel() {
+
+    init {
+        // TODO 추후 서버 통신
+        _uiState.value = dummyUiState
+    }
 
     fun putTodo() = viewModelScope.launch {
+        // TODO 추후 서버 통신
         delay(500L)
-        // TODO 서버 통신
     }
-
-    fun checkUser(userIdx: Int) {
-        _uiState.update { uiState ->
-            val newToDoUsers = uiState.todoUsers.mapIndexed { idx, toDoUser ->
-                if (idx == userIdx) {
-                    if (toDoUser.isChecked) {
-                        return@mapIndexed toDoUser.copy(
-                            isChecked = false,
-                            dayOfWeeks = List(7) { false }
-                        )
-                    } else {
-                        return@mapIndexed toDoUser.copy(isChecked = true)
-                    }
-                }
-                toDoUser
-            }
-            uiState.copy(
-                todoUsers = newToDoUsers,
-                selectedUsers = newToDoUsers.filter { it.isChecked }
-            )
-        }
-    }
-
-    fun selectTodoDay(userIdx: Int, dayIdx: Int) {
-        val dayOfWeeks: List<Boolean> =
-            uiState.value.todoUsers[userIdx].dayOfWeeks.mapIndexed { idx, checkState ->
-                if (idx == dayIdx) {
-                    return@mapIndexed !checkState
-                }
-                checkState
-            }
-        _uiState.update { uiState ->
-            val newToDoUsers = uiState.todoUsers.mapIndexed { idx, toDoUser ->
-                if (idx == userIdx) {
-                    return@mapIndexed uiState.todoUsers[idx].copy(
-                        dayOfWeeks = dayOfWeeks
-                    )
-                }
-                toDoUser
-            }
-            uiState.copy(
-                todoUsers = newToDoUsers,
-                selectedUsers = newToDoUsers.filter { it.isChecked }
-            )
-        }
-    }
-
-    fun changeNotificationState() {
-        _uiState.update { uiState ->
-            uiState.copy(
-                isPushNotification = !uiState.isPushNotification
-            )
-        }
-    }
-
-    fun isActiveAddButton() = (uiState.value.buttonState == ButtonState.ACTIVE)
-    fun isBlankToDoName(): Boolean = todoName.value.isBlank()
-    fun setToDoNameState(isBlank: Boolean) {
-        _uiState.update { uiState ->
-            uiState.copy(isBlankTodoName = isBlank)
-        }
-    }
-
-    data class AddToDoUiState(
-        override val selectedUsers: List<ToDoUser> = emptyList(),
-        override val todoUsers: List<ToDoUser> = emptyList(),
-        override val isPushNotification: Boolean = false,
-        override val buttonState: ButtonState = ButtonState.INACTIVE,
-        override val isBlankTodoName: Boolean = true
-    ) : ToDoUiState()
 
     companion object {
-        private val dummyUiState = AddToDoUiState(
+        private val dummyUiState = UpdateToDoUiState(
             todoUsers = listOf(
                 ToDoUser().copy(
                     homyType = HomyType.GREEN,
